Add initial consensus model for multiple partition level support
Partition level smaller than a database can be catastrophic when replica set for certain records are unavailable.

It also leads to complicated consensus impl for transaction that take from different tables or records. You would need a coordinated commit that may be multi-step

It is preferable for a replica to contain the entire database.

It also simplifies consensus for transactions. Makes it possible to implement Raft or Paxos