feat: implement statements parsing
This should round up current work on the parser. Phew! Such a slugfest!

Allow `exist` where `exists` can be used.

Enforcing reserved keywords. Before zeroql wasn't going to have reserved keywords cos it felt cool and surrealql does that too, however I don't think this property mix well with SQL keyword permutation property. It is easy to have a typo be interpreted as something else and I got bit by this while implementing the parser. For example, `DEFINE  NAMESPACE IF EXISTS ns` is not conceptually valid and this might have been a typo, but the parser goes ahead and interpretes `IF` as an identifier